Better implementation of checks
Remove duplicate code and transfer code for doing permission checks to utils folder. To utilize checks in cogs, import checks from .utils
